backend/requirements.txt
# Web Framework
fastapi>=0.109.0
hypercorn>=0.16.0
starlette>=0.35.0

# Database
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
pwdlib[argon2]
bcrypt>=4.0.1

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
aiohttp>=3.9.0
httpx>=0.26.0

# RSS
feedparser>=6.0.10
feedgenerator>=2.1.0

# AI/LLM
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-ollama>=0.1.0
langchain-community>=0.0.10

# Scheduling
APScheduler>=3.10.4

# Logging
loguru>=0.7.2

# Zotero
pyzotero>=1.5.0

# ArXiv
arxiv>=2.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
